refactor: extract handle click outside logic to hook
